requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-ckeditor==6.0.0
django-haystack==3.0
django-js-asset==1.2.2
django-redis==4.12.1
docopt==0.6.2
idna==2.10
importlib-metadata==3.7.3
jieba==0.42.1
Markdown==3.3.4
pipreqs==0.4.10
PyMySQL==1.0.2
pytz==2021.1
redis==3.5.3
requests==2.25.1
setuptools-scm==6.0.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
uWSGI==2.0.19.1
Whoosh==2.7.4
yarg==0.1.9
zipp==3.4.1
